requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
python-dotenv==1.0.1
intervaltree==3.1.0
pytest==8.3.3